requirements.txt
streamlit
python-dotenv
groq
pymupdf
pillow
streamlit-pdf